+++ data/response/rentalInformationResponse.go
@@ -4,7 +4,7 @@ type RentalInformationResponse struct {
 	ID           uint                 `json:"id"`
 	Name         string               `json:"name"`
 	Bedrooms     uint                 `json:"bedrooms"`
-	Bathrooms    uint                 `json:"bathrooms"`
+	Bathrooms    float64              `json:"bathrooms"`
 	Description  string               `json:"description"`
 	Location     LocationResponse     `json:"location"`
 	RentalStatus RentalStatusResponse `json:"rentalStatus"`
